requirement.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.4
django-import-export==4.0.0
django-jazzmin==3.0.0
django-taggit==3.0.0
mysqlclient==2.2.4
Pillow==9.1.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
